Update patch set 1
Patch Set 1:

(1 comment)

When a user signs up, do we know that the email they put in the box is the email that is stored? If the email is taken directly from Jira, then we probably shouldn't present the user with the option to enter in a different email, because it doesn't seem to be respected.

Also, less importantly, "Email addr" probably isn't the best label for that field. But I cannot see where that comes from in the code

Patch-set: 1
Reviewer: Gerrit User 6929 <6929@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0