Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > When a user signs up, do we know that the email they put in the box
 > is the email that is stored? If the email is taken directly from
 > Jira, then we probably shouldn't present the user with the option
 > to enter in a different email, because it doesn't seem to be
 > respected.
 > 
 > Also, less importantly, "Email addr" probably isn't the best label
 > for that field. But I cannot see where that comes from in the code

On the first part: tried a freshly cloned install with changes applied and it does not seem to display this issue. During initial development of this that was an issue, but it was resolved before pushing changes.

On the second part: email as a term was tested, but when implemented in the userprofile model as a mutable field appeared to conflict with the immutable (or rather difficult to access) field in the user model defined by Django. Any suggestions for alternative names for that field are welcome.

Patch-set: 2
Reviewer: Gerrit User 7081 <7081@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0