Update patch set 1
Patch Set 1:

@Fatih
I share your annoyance, but this is a limitation of the stackstorm framework. All action and workflow definitions ( actions/*.yaml ) must be in the actions directory. So if we created a 'fuel' subdirectory, we would only be able to put one file into it.
We could create a single actions/src directory and throw all the implementation files (*.sh or *.py) into there. Thoughts?

Patch-set: 1