Update patch set 1
Patch Set 1:

No problem. :)

I wouldn't say anger but while looking at the patches, I noticed it and thought it can be structured a bit better but I wasn't aware of the limitation.

This is definitely not the most crucial thing about the work you are doing so I am fine with whatever you do about my earlier comment; leave as is or move to src folder.

Patch-set: 1