Update patch set 2
Patch Set 2:

> (1 comment)

It doesn't cause any issues, and it doesn't require too much testing as this model was added just a couple commits prior. Having the primary handle for labs be their name instead of a numerical ID just makes it easier to reference a specific lab when creating or working with resources, especially through the API. As it's a primary key now, that is what is used in the API when serializing the object.

Since it is explicitly unique, no lab can be created with the same name as another so it can still be used as a unique ID.

Patch-set: 2