Update patch set 2
Patch Set 2:

> Patch Set 2: Published edit on patch set 1.

Patch set 2 allows the use of HTTPS basic authentication to directly use get and post requests on the API. Previously, token authentication did not work correctly to access the API. There seems to be some link that's missing between api tokens and users themselves that would be resolved outside of the API itself and within the 'account' app. For this reason, the implementation of token based authentication is still here but can not yet be used.

Patch-set: 2