Move Authentication to Linux Foundation Identity
Switch the user authentication source from using OAuth with JIRA to CAS
with the Linux Foundation Identity system.

This removes the need to maintain a private key file for connecting to
the authentication service, and allows us to no longer worry about user
passwords.

There are a couple changes still needed for this:
 - The development docker-compose file (still to come) should have a CAS
   instance brought up so real accounts do not need to be created.
   (ex: apereo/cas:v5.3.2)
 - Migrations for removing passwords from the Users table.
 - Removing the ability to modify fields which come from identity
   (email, timezone, name, etc)
 - The UI needs to be changed from saying "Login with JIRA" to just
   "Login"

A couple things that need to be decided:
 - What to do about scheduling/requesting access? There is code for
   creating JIRA tickets still here that either needs to be removed,
   with an alternative approach found, or we setup a JIRA account and
   auth just for those changes.
 - Superuser/Admin access - As this change allows us to pull in group
   information as well, should this access be granted through a specific
   group (opnfv-laas-admins) or continue being managed through the UI?

Note: The change from 'os.environ' to 'os.environ.get' allows us to
verify the settings.py file (by running python
src/pharos_dashboard/settings.py) and will be moved to another commit.

Change-Id: I8f6930eaa22dfb1790d2b53c7b0d08d9d181bbdb
JIRA: PHAROS-405
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>