Adds Logging
This commit adds a logging framework to follow a user session and
adds a bunch of useful logging. This commit also includes a couple of
bug fixes that rely on the logging features.

Log files are now put into /var/log/pharos-dashboard by default
and printed to stdout. A docker volume is mounted at the default
log dir.

Change-Id: Ic79db4bb89c25c3143428f9ff56eae3a0b937eee
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>