Update patch set 4
Patch Set 4:

> (3 comments)
 > 
 > No real bugs are immediately visible, but these should probably be
 > addressed before merging.

will fix whitespace.

On completed vs fulfilled vs deployed vs ended:
There are two things here. First, I think the word choice 'completed' makes sense here as this is the content of the email that gets sent to the user, not some method signature or something we as the developers may get confused on. Also, symantically, I think 'completed' was an unfortunate word choice from the beginning to represent when a job is over precisely because of this confusion between when it is fulfilled (all the work to deploy it is over) and when the job is over (the booking expired and is cleaned up). I will edit the email to try and be more clear.

And I don't think it will be really worth it to implement a new code path and extra HTTP request, etc etc, just for an empty message. I think the preview pane being empty is itself enough of an indication to the user that no message is selected. If we could solve this just in the template (have the iframe contain some "no message selected" doc by default that gets overridden) then that would be good.

Patch-set: 4
Reviewer: Gerrit User 6929 <6929@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0