Fixed PTL Checks
First, this commit correctly parses the gerrit urls and can
redirect itself to the raw info file if the user provides
the html rendered one.
This commit also fixes the way it checks the user against the
found PTLs. The returned data changed format when support was added for
sub-PTLs and the checks were never updated.

Change-Id: I5801d207ec823c718a5beb5d5e78ee5780ec65e0
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>