Update patch set 4
Patch Set 4:

> (2 comments)

The bootstrap version is what is used on the booking detail page for hiding and showing each panel

For the second part, probably doesn't matter much, but say only half the elements have loaded and the hide/show function is called. It will toggle the visibility of half of the elements, and then the visibility of all affected elements will be out of phase by 1 click cycle. Just having a variable at the top level of the javascript block that you toggle would work around this, as it would store the visibility state of all the elements rather than each element remembering it itself.

Patch-set: 4
Reviewer: Gerrit User 7081 <7081@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0