Update patch set 4
Patch Set 4:

> (1 comment)

I did it the way I did to reduce verbosity actually, as this gives me a way of indexing into the dictionary with the type of each task itself, and keeping the named empty lists lets me access them independently later. If it would be more clear what's going on to use a set of if statements in the for loop I can do that, it just seemed simpler while I was making it to do it this way.

Patch-set: 4
Reviewer: Gerrit User 7081 <7081@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0