Adds pdf and idf into api
The Pod Descriptor File (pdf) and Installer descriptor file (idf)
are now hosted in the api. The url endpoint where the lab
can retrieve them are now part of a software task.

An OPNFV task also contains a new dictionary that describes how bridges
should be configured on the jumphost. This information is not
contained in the pdf/idf but is needed by the lab.

Change-Id: I6971279979ba180725926035bd9db481aafb1073
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>