Update patch set 7
Patch Set 7:

(4 comments)

Using the validity structure helps simplify doing checks: the step that knows the most about the validity of it's data (the step itself) has the responsibility to mark whether the data can be safely used. This way we avoid just assuming existence implies integrity for under construction models. It isn't a huge deal so far for this workflow, but it's definitely helpful for me in making sure nothing falls apart because of me forgetting to do some integrity check in an accessing step.

Also, sshfs didn't seem to sync before I committed last so there were a couple new changes that didn't make it in to the last patch that will be there when I push up in a little bit.

Patch-set: 7
Label: Verified=0